requirements.txt
diffusers>=0.29.0
torch
torchvision
Pillow
numpy
jupyter